Add unitConversion constant to helpers
