+++ packages/lib-classifier/src/helpers/index.js
@@ -6,6 +6,7 @@ export { default as roundCoordinates } from './roundCoordinates'
 export { default as shownMarks } from './shownMarks'
 export { default as subjectsSeenThisSession } from './subjectsSeenThisSession'
 export { default as subjectViewers } from './subjectViewers'
+export { default as unitConversions } from './unitConversions'
 export { default as validateMimeType } from './validateMimeType'
 export { default as withFeatureFlag } from './withFeatureFlag'
 export { default as withStores } from './withStores'
